requirements.txt
SQLAlchemy~=2.0.41
the-new-hotness~=1.3.0
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.15
openai==0.27.8
venom-bot==3.0.2
pydantic==1.10.7
python-multipart==0.0.6
aiosqlite==0.18.0
loguru==0.7.0


sqlalchemy[asyncio]==2.0.23
asyncmy==0.2.9
python-dotenv==1.0.1
